Start cleaning up tests/ suite. Rename python test modules to be discoverable by pytest. Replace incomplete tests.testPopulate.test_double_sided_exp_scale with deterministic tests test_double_sided_exp_negative_sign and test_double_sided_exp_positive_sign and make consistent with method moving to psrpoppy/galactic_ops.py. Remove pointless path appending and shebang.
